Change initial matrix type from int to float

+++ src/DCT.h
@@ -2,7 +2,7 @@
 #define DCT_H
 
 float cos_DCT(int sumIndex, int DCTArrayIndex, int numOfElement);
-float adderFunction(int *imageMatrix, int DCTArrayIndex, int numOfElement);
-void DCT_Function(float *DCTArray, int *imageMatrix, int numOfElement);
+float adderFunction(float *imageMatrix, int DCTArrayIndex, int numOfElement);
+void DCT_Function(float *DCTArray, float *imageMatrix, int numOfElement);
 
 #endif // DCT_H
